We don't want prettier or prettierd installed globally, because then we can't opt out of formatting in non-prettier enabled projects.
The files that activate prettier/prettierd include package.json, which
is all npm projects. This is somewhat unfortunate for folks who work in
many different projects.

It _does_ make sense to install prettier in a project if you want to use
prettier in that project, as its the only way to lint/format your
project.  The advantage with prettierd, however, can never be realized,
but maybe the extra memory tradeoff is worth it. I have enough RAM, so 🤷

See:
- https://github.com/stevearc/conform.nvim/issues/295#issuecomment-1972530238
- https://github.com/stevearc/conform.nvim/pull/315
- https://github.com/stevearc/conform.nvim/issues/268

Thanks @stevearc for helping me work through this issue.